added global exception handles and also resolved the pom files for version

+++ Create-Task-Application/src/main/java/com/batch5/Create_Task_Application/UserModule/service/UserRoleService.java
@@ -10,7 +10,7 @@ public interface UserRoleService {
 
     List<UserRole> getAllRoles();
 
-    UserRole getRoleById(Long roleId);
+    UserRole getRoleById(Integer roleId);
 
-    void deleteRole(Long roleId);
+    void deleteRole(Integer roleId);
 }
